Moved MutationsJSON.java to service layer
