+++ business/src/main/resources/applicationContext-business.xml
@@ -48,7 +48,6 @@
         <!-- enable component scanning (beware that this does not enable mapper scanning!) -->
         <context:component-scan base-package="org.mskcc.cbio.portal.service" />
          <context:component-scan base-package="org.cbioportal.persistence" />
-         <context:component-scan base-package="org.cbioportal.service" />
 
         <!-- enable autowire -->
         <context:annotation-config />
