+++ core/pom.xml
@@ -17,11 +17,6 @@
       <artifactId>business</artifactId>
       <version>${project.version}</version>
     </dependency>
-    <dependency>
-      <groupId>org.mskcc.cbio</groupId>
-      <artifactId>persistence-mybatis</artifactId>
-      <version>${project.version}</version>
-    </dependency>
 
     <!-- antisamy -->
     <dependency>
