+++ portal/src/main/webapp/js/api/cbioportal-datamanager.js
@@ -459,10 +459,9 @@ window.initDatamanager = function (genetic_profile_ids, oql_query, cancer_study_
 	    var num_mutations_promise = new $.Deferred();
 	    special_sample_data_promises.push(num_mutations_promise.promise());
 	    clinicalMutationColl.fetch({
-		type: "POST",
+		type: "GET",
 		data: {
 		    mutation_profile: self.getMutationProfileId(),
-		    cmd: "count_mutations",
 		    case_ids: self.getSampleIds().join(" ")
 		},
 		success: function (response) {
