+++ service/pom.xml
@@ -10,11 +10,16 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>service</artifactId>
-
+    
     <dependencies>
         <dependency>
             <groupId>org.mskcc.cbio</groupId>
             <artifactId>persistence-mybatis</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.mskcc.cbio</groupId>
+            <artifactId>core</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
