+++ service/src/main/java/org/cbioportal/service/MutationService.java
@@ -2,14 +2,24 @@ package org.cbioportal.service;
 
 import org.cbioportal.model.Mutation;
 import org.cbioportal.persistence.dto.AltCount;
+import org.mskcc.cbio.portal.dao.DaoException;
 
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 public interface MutationService {
 
     List<Mutation> getMutationsDetailed(List<String> geneticProfileStableIds, List<String> hugoGeneSymbols,
                                         List<String> sampleStableIds, String sampleListStableId);
 
-    List<AltCount> getMutationsCounts(String type, String hugoGeneSymbol, Integer start, Integer end,
-                                      List<String> cancerStudyIdentifiers, Boolean perStudy);
+    Map<String,List> getMutationMatrix(List<String> sampleStableIds, String mutationGeneticProfileStableId,
+                                       String mrnaGeneticProfileStableId, String cnaGeneticProfileStableId,
+                                       String drugType) throws IOException, DaoException;
+
+    Map<String, Integer> getMutationCount(String mutationGeneticProfileStableId, String sampleStableIds);
+
+    List<Map<String,Object>> getSmg(String mutationGeneticProfileStableId, String sampleStableIds) throws DaoException;
 }
